requirements.txt
langchain-openai==0.0.1
langchain-core==0.0.1
langchain-pinecone==0.0.1
streamlit==1.11.0
pinecone-client==2.2.1
python-dotenv==0.21.0